Refactoring (ViewModels, Autommaper)

+++ Api_Persona/Persona.Entidades/Pais.cs
@@ -8,7 +8,7 @@ namespace Persona.Entidades
     public class Pais
     {
         [Key]
-        public Int16 Id { get; set; }
+        public int Id { get; set; }
         public string Descripcion { get; set; }
     }
 }
